Move these api tests
